// lib/interfaces/application/applicationResponse.ts
// Generated by https://quicktype.io

export interface ApplicationResponse {
  results: Application;
}

export interface Application {
  user_id: string;
  legal_first_names: string;
  legal_last_names: string;
  nationality: string;
  email: string;
  phone: string;
  date_of_birth: string | number;
  gender: string;
  passport_number: string;
  passport_expiration_date: string | number;
  residence: string;
  residence_address: string;
  job: string;
  comments: string;
  status: 'draft' | 'confirmed';
  created_at: string;
  updated_at: string;
  photos: any[];
  documents: any[];
  payments: any[];
}